Rearranged the Src\ directory and added express includes such as Core\Core.h and Render\Render.h to include everything in the sub-directory.

+++ Src/Render/Circle.cpp
@@ -13,7 +13,7 @@ namespace glz
 	{
 		clearParts();
 
-		Float theta = 2.0f * glz::PI / (Float)segments;
+		Float theta = 2.0f * (float)glz::PI / (Float)segments;
 		Float cs = std::cosf(theta);
 		Float sn = std::sinf(theta);
 		Float t = 0.0f;
